Solution to Calculate Area and Perimeter of Rectangles

+++ Calculate Area and Perimeter of Rectangles@@67877d61dc107486ad5c274b/code.c
@@ -13,6 +13,6 @@ int main(){
     }
     for(int i=0;i<num;i++){
         printf("Rectagle %d: Area = %.2f, Perimeter = %.2f",i+1,
-        (ap[i].lenght)*(ap[i].breadth),2*(ap[i].lenght+ap[i].breadth));
+        (float)(ap[i].lenght)*(ap[i].breadth),2*(ap[i].lenght+ap[i].breadth));
     }
 }
